feat(sdr-store): refactor file storage system and add file cache
This commit was a refactoring of the file storage system from Mavlink and the inclusion of a new file cache system. Firstly, file types related to ASV SDR have been moved to their own classes to make it clearer where those classes should be used.

The introduction of a file cache system comes with many advantages. By reusing the same file set, we can reduce the overhead of creating new file objects and discarding old ones. This change could potentially increase the performance of the system by lessening the number of file read/writes that need to take place.

Additionally, File Rename and Move operations became safer as files that are currently in use will not be processed. This ensures the integrity of file operation processes.

Lastly, a variety of useful logging information was added for debugging and traceability purposes. It offers easier tracking of events such as renaming, moving, or deleting files.

+++ src/Asv.Mavlink/Tools/ListDataFile/ListDataStore.cs
@@ -1,11 +1,13 @@
 #nullable enable
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
+using System.Reactive.Linq;
+using System.Threading;
 using Asv.Common;
 using Asv.IO;
 using NLog;
@@ -13,19 +15,23 @@ using NLog;
 namespace Asv.Mavlink;
 
 public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDataStore<TMetadata, TKey> 
-    where TMetadata : ISizedSpanSerializable, new() where TKey : notnull
+    where TMetadata : ISizedSpanSerializable, new() 
+    where TKey : notnull
 {
     private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
     private readonly string _rootFolder;
     private readonly ListDataStoreFormat<TKey,TMetadata> _keyConverter;
     private readonly ListDataFileFormat _fileFormat;
+    private readonly TimeSpan _fileCacheTime;
     private readonly Dictionary<TKey,ListDataStoreEntry<TKey>> _entries;
     private readonly object _sync = new();
     private readonly RxValue<ushort> _count;
     private readonly RxValue<ulong> _size;
-    
+    private readonly ReaderWriterLockSlim _fileCacheLockSlim = new();
+    private readonly List<CachedFileWithRefCounter<TMetadata,TKey>> _fileCache = new();
+    private int _checkCacheInProgress;
 
-    public ListDataStore(string rootFolder, ListDataStoreFormat<TKey,TMetadata> keyConverter, ListDataFileFormat fileFormat)
+    public ListDataStore(string rootFolder, ListDataStoreFormat<TKey,TMetadata> keyConverter, ListDataFileFormat fileFormat, TimeSpan fileCacheTime)
     {
         if (string.IsNullOrEmpty(rootFolder))
             throw new ArgumentException("Value cannot be null or empty.", nameof(rootFolder));
@@ -33,6 +39,7 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
         _keyConverter = keyConverter ?? throw new ArgumentNullException(nameof(keyConverter));
         _keyConverter.DisposeItWith(Disposable);
         _fileFormat = fileFormat ?? throw new ArgumentNullException(nameof(fileFormat));
+        _fileCacheTime = fileCacheTime;
         _entries = new Dictionary<TKey, ListDataStoreEntry<TKey>>(keyConverter.KeyComparer);
         _count = new RxValue<ushort>().DisposeItWith(Disposable);
         _size = new RxValue<ulong>().DisposeItWith(Disposable);
@@ -41,8 +48,13 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
             Directory.CreateDirectory(_rootFolder);
         }
         InternalUpdateEntries();
+        Disposable.AddAction(ClearFileCache);
+        Observable.Timer(_fileCacheTime, _fileCacheTime).Subscribe(CheckFileCacheForRottenFiles)
+            .DisposeItWith(Disposable);
     }
-    
+
+   
+
     public IEnumerable<IListDataStoreEntry<TKey>> GetEntries()
     {
         lock (_sync)
@@ -56,6 +68,7 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     
     private void InternalUpdateEntries()
     {
+        Logger.Trace("Update entries");
         _entries.Clear();
         foreach (var entry in InternalGetEntries(null))
         {
@@ -68,10 +81,12 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         _count.Value = (ushort)_entries.Count(x => x.Value.Type == StoreEntryType.File);
         _size.Value = (ulong)_entries.Where(x => x.Value.Type == StoreEntryType.File).Sum(x => new FileInfo(x.Value.FullPath).Length);
+        Logger.Info($"Update statistics: rec:{_count.Value}, size:{_size.Value}");
     }
 
     private IEnumerable<ListDataStoreEntry<TKey>> InternalGetEntries(ListDataStoreEntry<TKey>? parentFolder)
     {
+        
         foreach (var directory in Directory.EnumerateDirectories(parentFolder?.FullPath ?? _rootFolder))
         {
             var folderInfo = new DirectoryInfo(directory);
@@ -121,12 +136,112 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
         }
     }
 
+    #region File cache
+
+    private void ClearFileCache()
+    {
+        _fileCacheLockSlim.EnterWriteLock();
+        try
+        {
+            foreach (var file in _fileCache)
+            {
+                file.ImmediateDispose();
+            }
+            _fileCache.Clear();
+        }
+        finally
+        {
+            _fileCacheLockSlim.ExitWriteLock();
+        }
+        _fileCacheLockSlim.Dispose();
+    }
+
+    private bool TryImmediatelyRemoveFromCache(TKey id)
+    {
+        try
+        {
+            _fileCacheLockSlim.EnterUpgradeableReadLock();
+            if (_fileCache.Count == 0) return true;
+            var item = _fileCache.FirstOrDefault(file => file.RefCount == 0 && file.Id.Equals(id));
+            if (item == null) return true;
+            
+            if (item.RefCount != 0) return false;
+            
+            _fileCacheLockSlim.EnterWriteLock();
+            Logger.Trace("Immediately remove file from cache: {0}", item.Id);
+            item.ImmediateDispose();
+            _fileCache.Remove(item);
+            _fileCacheLockSlim.ExitWriteLock();
+            return true;
+        }
+        finally
+        {
+            _fileCacheLockSlim.ExitUpgradeableReadLock();
+        }
+    }
     
+    private CachedFileWithRefCounter<TMetadata,TKey> GetOrAddFileToCache(TKey id, Func<TKey,ListDataFile<TMetadata>> add)
+    {
+        try
+        {
+            _fileCacheLockSlim.EnterUpgradeableReadLock();
+            var exist = _fileCache.FirstOrDefault(_ => _.Id.Equals(id));
+            if (exist != null)
+            {
+                Logger.Trace($"Return file '{id}' from cache");
+                exist.AddRef();
+                return exist;
+            }
+            else
+            {
+                _fileCacheLockSlim.EnterWriteLock();
+                var file = add(id);
+                var wrapper = new CachedFileWithRefCounter<TMetadata, TKey>(file, id);
+                wrapper.AddRef();
+                Logger.Trace($"Add file '{id}' to cache");
+                _fileCache.Add(wrapper);
+                _fileCacheLockSlim.ExitWriteLock();
+                return wrapper;
+            }
+        }
+        finally
+        {
+            _fileCacheLockSlim.ExitUpgradeableReadLock();
+        }
+    }
+    
+    private void CheckFileCacheForRottenFiles(long delay)
+    {
+        if (Interlocked.CompareExchange(ref _checkCacheInProgress,1,0) != 0) return;
+        try
+        {
+            _fileCacheLockSlim.EnterUpgradeableReadLock();
+            if (_fileCache.Count == 0) return;
+            var list = _fileCache.Where(file=>file.RefCount == 0).Where(file => DateTime.Now - file.DeadTimeBegin > _fileCacheTime).ToImmutableList();
+            if (list.Count == 0) return;
+            _fileCacheLockSlim.EnterWriteLock();
+            foreach (var file in list)
+            {
+                file.ImmediateDispose();
+                _fileCache.Remove(file);
+                Logger.Trace("Remove file from cache: {0}", file.Id);
+            }
+            _fileCacheLockSlim.ExitWriteLock();
+        }
+        finally
+        {
+            _fileCacheLockSlim.ExitUpgradeableReadLock();
+            Interlocked.Exchange(ref _checkCacheInProgress, 0);
+        }
+    }
+    
+    #endregion
 
     public bool MoveFolder(TKey id, TKey newParentId)
     {
         lock (_sync)
         {
+            Logger.Info("Move folder '{0}' to '{1}'", id, newParentId);
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.Folder) return false;
             if (_entries.TryGetValue(newParentId, out var newParent) == false) return false;
@@ -186,6 +301,7 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
 
             if (_keyConverter.TryGetFolderKey(folderInfo, out var newFolderId) == false)
             {
+                Logger.Error("Couldn't create folder {0}", name);
                 throw new InvalidOperationException($"Couldn't create folder {name}");
             }
             var newEntry = new ListDataStoreEntry<TKey>
@@ -199,9 +315,10 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
             if (_entries.Where(x => _keyConverter.KeyComparer.Equals(x.Value.ParentId, parentId))
                 .Any(x => x.Value.Name == name))
             {
+                Logger.Error("Folder {0} already exist",name);
                 throw new ListDataFolderAlreadyExistException(name);
             }
-            
+            Logger.Info("Create folder {0}",name);
             _entries.Add(newFolderId,newEntry);
             Directory.CreateDirectory(newFolderPath);
             
@@ -214,12 +331,42 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         lock (_sync)
         {
+            Logger.Info("Delete folder {0}",id);
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.Folder) return false;
+            var files = InternalGetSubFiles(id);
+            foreach (var file in files)
+            {
+                if (TryImmediatelyRemoveFromCache(file) == false)
+                {
+                    throw new Exception("Can't rename folder: file in folder currently in use");
+                }
+            }
             Directory.Delete(entry.FullPath,true);
             InternalUpdateEntries();
             return true;
         }
+    }
+
+    private IEnumerable<TKey> InternalGetSubFiles(TKey folderId)
+    {
+        if (_entries.TryGetValue(folderId, out var folder) == false) yield break;
+        foreach (var item in _entries)
+        {
+            if (item.Value.ParentId.Equals(folder.Id) == false) continue;
+            if (item.Value.Type == StoreEntryType.File)
+            {
+                yield return item.Value.Id;
+            }
+            else
+            {
+                foreach (var file in InternalGetSubFiles(item.Value.Id))
+                {
+                    yield return file;
+                }
+            }
+        }
+        
     }
 
     public bool ExistFolder(TKey id)
@@ -236,10 +383,19 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         lock (_sync)
         {
+            Logger.Info($"Rename folder {id} to {newName}");
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.Folder) return false;
             var rootFolder = Path.GetDirectoryName(entry.FullPath);
             var newFolderName = rootFolder != null ? Path.Combine(rootFolder, newName) : newName;
+            var files = InternalGetSubFiles(id);
+            foreach (var file in files)
+            {
+                if (TryImmediatelyRemoveFromCache(file))
+                {
+                    throw new Exception("Can't rename folder: file in folder currently in use");
+                }
+            }
             Directory.Move(entry.FullPath,newFolderName);
             InternalUpdateEntries();
             return true;
@@ -268,31 +424,36 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
         }
     }
 
-    public IListDataFile<TMetadata> Open(TKey id)
+    public ICachedFileWithRefCounter<TMetadata> Open(TKey id)
     {
         lock (_sync)
         {
             if (_entries.TryGetValue(id, out var entry) == false)
             {
+                Logger.Error("File with [{0}] not found", id);
                 throw new FileNotFoundException($"File with [{id}] not found");
             }
             
             var fileInfo =  new FileInfo(entry.FullPath);
             if (fileInfo.Exists == false)
             {
+                Logger.Error("File with [{0}]{1} not found", id, fileInfo.FullName);
                 throw new FileNotFoundException($"File with [{id}]{fileInfo.FullName} not found", fileInfo.FullName);
             }
-            var file = new ListDataFile<TMetadata>(File.Open(fileInfo.FullName, FileMode.OpenOrCreate, FileAccess.ReadWrite), _fileFormat, true);
-            return file;
+
+            return GetOrAddFileToCache(id,
+                _ => new ListDataFile<TMetadata>(
+                    File.Open(fileInfo.FullName, FileMode.OpenOrCreate, FileAccess.ReadWrite), _fileFormat, true));
         }
     }
 
-    public IListDataFile<TMetadata> Create(TKey id, TKey parentId, Action<TMetadata> defaultMetadata)
+    public ICachedFileWithRefCounter<TMetadata> Create(TKey id, TKey parentId, Action<TMetadata> defaultMetadata)
     {
         lock (_sync)
         {
             if (_entries.ContainsKey(id))
             {
+                Logger.Error("Entry {0} already exist",id);
                 throw new Exception($"Entry {id} already exist");
             }
             
@@ -302,10 +463,16 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
                 parentId = _keyConverter.RootFolderId;
                 rootFolder = _rootFolder;
             }
+            else
+            {
+                rootFolder = parent.FullPath;
+            }
+                
             var name = _keyConverter.GetFileNameByKey(id);
             var fileInfo = new FileInfo(Path.Combine(rootFolder, name));
             if (fileInfo.Exists)
             {
+                Logger.Error("File [{0}]{1} already exist", id, fileInfo.FullName);
                 throw new Exception($"File [{id}]{fileInfo.FullName} already exist");
             }
             var file = new ListDataFile<TMetadata>(File.Open(fileInfo.FullName, FileMode.OpenOrCreate, FileAccess.ReadWrite), _fileFormat, true);
@@ -318,7 +485,7 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
                 ParentId = parentId,
                 FullPath = fileInfo.FullName,
             });
-            return file;
+            return GetOrAddFileToCache(id,_=>file);
         }
     }
 
@@ -326,8 +493,11 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         lock (_sync)
         {   
+            Logger.Info("Delete file {0}",id);
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.File) return false;
+            // we need to remove file from cache, if we want to modify it
+            if (TryImmediatelyRemoveFromCache(id) == false) throw new Exception("Can't delete file: it is currently in use");
             File.Delete(entry.FullPath);
             _entries.Remove(id);
             return true;
@@ -346,10 +516,13 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         lock (_sync)
         {
+            Logger.Info("Rename file {0} to {1}",id,newName);
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.File) return false;
             var rootFolder = Path.GetDirectoryName(entry.FullPath);
             var newFileName = rootFolder != null ? Path.Combine(rootFolder, newName) : newName;
+            // we need to remove file from cache, if we want to modify it
+            if (TryImmediatelyRemoveFromCache(id) == false) throw new Exception("Can't rename file: it is currently in use");
             File.Move(entry.FullPath,newFileName);
             entry.Name = newName;
             entry.FullPath = newFileName;
@@ -361,11 +534,14 @@ public class ListDataStore<TMetadata, TKey> : DisposableOnceWithCancel, IListDat
     {
         lock (_sync)
         {
+            Logger.Info("Move file {0} to {1}",id,newParentId);
             if (_entries.TryGetValue(id, out var entry) == false) return false;
             if (entry.Type != StoreEntryType.File) return false;
             if (_entries.TryGetValue(newParentId, out var newParent) == false) return false;
             if (newParent.Type != StoreEntryType.Folder) return false;
             var newFileName = Path.Combine(newParent.FullPath, entry.Name);
+            // we need to remove file from cache, if we want to modify it
+            if (TryImmediatelyRemoveFromCache(id) == false) throw new Exception("Can't move file: it is currently in use");
             File.Move(entry.FullPath,newFileName);
             entry.FullPath = newFileName;
             entry.ParentId = newParentId;
@@ -384,3 +560,4 @@ public class ListDataStoreEntry<TKey> : IListDataStoreEntry<TKey>
     public TKey ParentId { get;set; }
     public string FullPath { get;set; }
 }
+
